Refactor chatwidget protocol flows into modules
